requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
coverage==7.6.0
cramjam==2.8.3
dateparser==0.7.6
distro==1.9.0
Django==5.0.7
django-cors-headers==4.4.0
django-rest-passwordreset==1.4.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
drf-yasg==1.21.7
exceptiongroup==1.2.2
# feedparser==6.0.11
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.0
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
jiter==0.5.0
load-dotenv==0.1.0
mock==5.1.0
multidict==6.0.5
numpy==1.26.4
openai==1.45.0
packaging==24.1
pillow==10.4.0
playwright==1.46.0
pluggy==1.5.0
pydantic==2.9.1
pydantic_core==2.23.3
pyee==11.1.0
pygooglenews==0.1.2
PyJWT==2.8.0
pymongo==4.8.0
pytest==8.3.1
pytest-base-url==2.1.0
pytest-playwright==0.5.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
python-snappy==0.7.2
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.1
text-unidecode==1.3
tokenizers==0.19.1
tomli==2.0.1
tqdm==4.66.4
transformers==4.42.4
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.2
winkerberos==0.11.0
yarl==1.9.4
